+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.5.4
 - disable validate_email module blacklist updater #17
+- SMTP envelope MAIL FROM address #16
 
 ## 1.5.3 (2021-08-02)
 - group syntax is no more reported as erroneous (ex: undisclosed-recipients:;)